Group No. 3/fraud/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2
requests==2.31.0